Fix yet another two bugs on array index checking
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@434367 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/nio/src/main/java/java/nio/CharSequenceAdapter.java
@@ -117,7 +117,7 @@ final class CharSequenceAdapter extends CharBuffer {
 	}
 
     public final CharBuffer put(char[] src, int off, int len) {
-        if ((off < 0 ) || (len < 0) || off + len > src.length) {
+        if ((off < 0 ) || (len < 0) || (long)off + (long)len > src.length) {
             throw new IndexOutOfBoundsException();
         }
